[Refactored]: Moved folders into pkg folder
