+++ pkg/controller/emailController.go
@@ -4,8 +4,8 @@ import (
 	"errors"
 	"github.com/gin-gonic/gin"
 	"net/http"
-	"se-school-case/dto"
-	"se-school-case/service"
+	"se-school-case/pkg/dto"
+	service2 "se-school-case/pkg/service"
 )
 
 func PostAddUserEmail(c *gin.Context) {
@@ -18,8 +18,8 @@ func PostAddUserEmail(c *gin.Context) {
 	}
 
 	// Handle email subscription in service layer
-	if err := service.HandleEmailSubscription(input.Email); err != nil {
-		if errors.Is(err, service.ErrEmailAlreadyExists) {
+	if err := service2.HandleEmailSubscription(input.Email); err != nil {
+		if errors.Is(err, service2.ErrEmailAlreadyExists) {
 			c.JSON(http.StatusConflict, gin.H{"error": "Email already exists"})
 		} else {
 			c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to add email"})
@@ -32,6 +32,6 @@ func PostAddUserEmail(c *gin.Context) {
 }
 
 func PostExplicitlyNotify(c *gin.Context) {
-	service.SendEmailNotificationsToAll()
+	service2.SendEmailNotificationsToAll()
 	c.JSON(http.StatusOK, gin.H{"message": "Successfully notified all users."})
 }
