+++ pkg/dto/emailSendDto.go
