+++ pkg/dto/rateAPIDto.go
