+++ pkg/model/rateModel.go
