+++ pkg/resource/commands.md
