+++ pkg/resource/email.html
