+++ pkg/show/data/email_notification_example.png
